\docType{class}
\name{bmProxy-class}
\alias{bmProxy-class}
\title{CLASS bmProxy}
\description{
bmProxy Class uses the RESTAPI to communicate with the
BrowserMob Proxy.
}
\details{
bmProxy is a generator object. To define a new bmProxy
class method `new` is called. The slots (default value)
that are user defined are:
}
\section{Slots}{
  \describe{ \item{\code{host}:}{Object of class
  \code{"character"}, giving the ip of the remote server.
  Defaults to localhost} \item{\code{port}:}{Object of
  class \code{"numeric"}, the port of the remote server on
  which to connect.} \item{\code{selHost}:}{Object of class
  \code{"character"}, giving the ip of the remote server.
  Defaults to localhost} \item{\code{selPort}:}{Object of
  class \code{"numeric"}, the port of the remote server on
  which to connect.} }
}
\examples{
\dontrun{
}
}

