\docType{class}
\name{errorHandle-class}
\alias{errorHandle-class}
\title{CLASS errorHandle}
\description{
class to handle errors
}
\details{
brief desc here
}
\section{Methods}{
  \describe{ \item{\code{new(...)}:}{ Create a new
  \code{errorHandle} object. ... is used to define the
  appropriate slots.} }
}
\examples{
\dontrun{
}
}

